Use EcoreUtil.copy(EObject) where possible for ICloneable.getCopy()
The old methods had to be updated each time a new attribute or feature was added to an EObject.
EcoreUtil.copy(EObject) copies everything so we don't need to maintain the classes.
We do need to de-reference some child objects and references, however, in the ICloneable.getCopy() implementations as we are creating a new copy.