Register additional IArchimatePackage Namespace URI
Register Namespace URI "http://www.bolton.ac.uk/archimate" in addition to "http://www.bolton.ac.uk/archimate/1.0"
In October 2010 I changed the Namespace URI from "http://www.bolton.ac.uk/archimate/1.0" to "http://www.bolton.ac.uk/archimate" in the ecore model but not in the plugin.xml.
It was only by good fortune that the ArchimatePackage was initialised before loading any model. This was done by the uk.ac.bolton.archimate.bizzdesign plug-in in a static initialiser of a class.
Once the uk.ac.bolton.archimate.bizzdesign plug-in is removed, initialisation of the package does not happen when models with the namespace URI "http://www.bolton.ac.uk/archimate" are loaded.
Also, show an error dialog if a package can't be loaded.