Add Label Provider extension point and refactor ImageFactory
Label Provider extension point allows plug-ins to contribute text labels and icon images to the UI.
Cases where an element's name is displayed now display a default name if the element's name is empty.
Icons are displayed in the menu items of the SearchWidget and the tables of Nested Relation dialogs.
Removed unused methods from ImageFactory.