Issue #123 - Shortcut/Link from view diagram elements to views
Added if statement to Model DND policies so that links to models can
be automatically formatted as mini-links. The code only performs the
following changes:

1) Positions the model reference at parent bottom right corner
   -25x, -25y
   
2) Sets the size of the model reference to 25x, 25y

The miniaturization only occurs if the following conditions are met:

1) Only a single model reference is being DND (no other objects)

2) The receiving parent object is a DiagramModelObject

It uses the current diagram elements, just resizes them to small
default size if conditions are met, so there should be minimal
regression risk.