Issue 132 LUNA correct git eol character
For this commit I have made no code changes whatsoever.
I used Eclipse Luna and followed the dev setup exactly.
The git repo was a new copy, rebased to archi-HEAD, and
hard reset, and cleaned.

Git was configued for Windows EOL (as specifed that it
should be) and was set to not to automatically change
any line endings.

git config core.eol crlf

git config core.autocrlf false

The repo still has a mix of LF and CRLF.

In this commit I have only done the following in Eclipse:

1. File > Convert Line Delimiters To > Windows

2. Filter Selections > *.java

3. Commit

I have not added anything to will automatically change LF
to CRLF in the future, so people are still free to commit
any line endings they want.