Fix circular bendpoints undo/redo not working
If a circular connection is made 3 bendpoints are also added.

If you then move one of these bendpoints, then undo this and undo the circular connection
and then redo this sequence, and undo it all again then the moved bendpoints are not removed.

This is because the undo/redo sequence in the sub-compound command was wrong.

Also, refactor this as the create bendpoints command is used more than once.