[Properties View] Workaround for "Copy" action still enabled when Properties View is closed
If the Properties View is open and no other View is providing a Tabbed Property View
then we see the default table in the Properties View and an enabled "copy" action.
If the Properties View is then closed the global "copy" action is still enabled and
will cause a "Widget is disposed" exception if selected. So we disable it.