[GEF] Fix possible SWTException "Widget is disposed" when displaying a Tooltip
If a workbench part is detached and a tooltip is initially displayed, when the part is re-attached the parent Shell is disposed which in turn disposes of the PopupHelper's Shell. This will lead to an SWT Error when the Shell is accessed again.

To fix this, create the PopupHelper Shell with the Control's Display instead of its Shell.

See https://github.com/eclipse/gef-classic/issues/469