[Themes] Use proper method to store theme font preference
We have to use this method for creating the font preference key for fonts so that the theme prefix is written and then it works with the high contrast theme.

Normally a font is stored in org.eclipse.ui.workbench.prefs like this:

  com.archimatetool.editor.MODEL_TREE_FONT=<fontdata>

and that works for all themes except the high contrast one.

The high contrast theme stores it like this:

  org.eclipse.ui.ide.systemDefault.com.archimatetool.editor.MODEL_TREE_FONT=<fontdata>